osa4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -41,6 +41,18 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  if (!user) {
+    return response.status(400).json({ error: 'userId missing or not valid' })
+  }
+  const addingUser = await Blog.findById(request.params.id)
+  if (addingUser.user.toString() !== user.id.toString()) {
+    return response.status(400).json({ error: 'not authorized to delete' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
